+++ src/main/java/plugily/projects/villagedefense/kits/level/MediumTankKit.java
@@ -41,7 +41,7 @@ public class MediumTankKit extends LevelKit {
     public MediumTankKit() {
         setName(getPlugin().getChatManager().colorMessage(Messages.KITS_MEDIUM_TANK_NAME));
         List<String> description = Utils.splitString(getPlugin().getChatManager().colorMessage(Messages.KITS_MEDIUM_TANK_DESCRIPTION), 40);
-        this.setDescription(description.toArray(new String[0]));
+        setDescription(description.toArray(new String[0]));
         setLevel(getKitsConfig().getInt("Required-Level.MediumTank"));
         KitRegistry.registerKit(this);
     }