+++ src/main/java/plugily/projects/villagedefense/kits/level/RunnerKit.java
@@ -45,7 +45,7 @@ public class RunnerKit extends LevelKit {
         setLevel(getKitsConfig().getInt("Required-Level.Runner"));
         setName(getPlugin().getChatManager().colorMessage(Messages.KITS_RUNNER_NAME));
         List<String> description = Utils.splitString(getPlugin().getChatManager().colorMessage(Messages.KITS_RUNNER_DESCRIPTION), 40);
-        this.setDescription(description.toArray(new String[0]));
+        setDescription(description.toArray(new String[0]));
         KitRegistry.registerKit(this);
     }