+++ src/main/java/plugily/projects/villagedefense/kits/premium/NakedKit.java
@@ -52,7 +52,7 @@ public class NakedKit extends PremiumKit implements Listener {
   public NakedKit() {
     setName(getPlugin().getChatManager().colorMessage(Messages.KITS_WILD_NAKED_NAME));
     List<String> description = Utils.splitString(getPlugin().getChatManager().colorMessage(Messages.KITS_WILD_NAKED_DESCRIPTION), 40);
-    this.setDescription(description.toArray(new String[0]));
+    setDescription(description.toArray(new String[0]));
     getPlugin().getServer().getPluginManager().registerEvents(this, getPlugin());
     KitRegistry.registerKit(this);
     setupArmorTypes();