+++ src/main/java/plugily/projects/villagedefense/kits/premium/TornadoKit.java
@@ -58,7 +58,7 @@ public class TornadoKit extends PremiumKit implements Listener {
   public TornadoKit() {
     setName(getPlugin().getChatManager().colorMessage(Messages.KITS_TORNADO_NAME));
     List<String> description = Utils.splitString(getPlugin().getChatManager().colorMessage(Messages.KITS_TORNADO_DESCRIPTION), 40);
-    this.setDescription(description.toArray(new String[0]));
+    setDescription(description.toArray(new String[0]));
     getPlugin().getServer().getPluginManager().registerEvents(this, getPlugin());
     KitRegistry.registerKit(this);
   }