+++ CHANGELOG.md
@@ -9,6 +9,9 @@ deploy webhooks to print changelog successfully
 
 That's all, matcher will stop when detects next line started with `###` match
 
+### 4.5.5 (30.12.2020)
+* Fixed Kit Menu does not work (NPE)
+
 ### 4.5.4 (09.12.2020 - 29.12.2020)
 * Fix NPE when kit name not found in language file
 * Now players can see their orbs in stats command